connect user and database

+++ pages/workouts/index.js
@@ -6,10 +6,13 @@ import { FavouriteButton } from "@/components/Workout";
 import { useState } from "react";
 import Login from "@/components/Login";
 import useSWR from "swr";
+import { useSession } from "next-auth/react";
 
 export default function WorkoutsPage() {
-  const { data: workouts = [], isLoading, mutate } = useSWR("/api/workouts");
   const { data: exercises = [] } = useSWR("/api/exercises");
+  const { data: session } = useSession();
+  const url = session ? `/api/workouts/user` : `/api/workouts`;
+  const { data: workouts = [], isLoading, mutate } = useSWR(url);
 
   const [isFavouritesMode, setisFavouritesMode] = useState(false);
 
@@ -18,7 +21,11 @@ export default function WorkoutsPage() {
   }
 
   async function handleDeleteWorkout(id) {
-    const response = await fetch(`/api/workouts/${id}`, {
+    if (!session) {
+      console.error("No active session found. Please log in.");
+      return;
+    }
+    const response = await fetch(`/api/workouts/user/${session.user.id}`, {
       method: "DELETE",
     });
     if (response.ok) {
@@ -27,10 +34,15 @@ export default function WorkoutsPage() {
   }
 
   async function handleToggleFavourite(id) {
+    if (!session) {
+      console.error("No active session found. Please log in.");
+      return;
+    }
+
     const workout = workouts.find((workout) => workout._id === id);
     if (workout) {
       workout.isFavourite = !workout.isFavourite;
-      const response = await fetch(`/api/workouts/${id}`, {
+      const response = await fetch(`/api/workouts/user/${session.user.id}`, {
         method: "PUT",
         headers: {
           "Content-Type": "application/json",
@@ -46,10 +58,6 @@ export default function WorkoutsPage() {
   const favouriteWorkouts = workouts.filter(
     (workout) => workout.isFavourite === true
   );
-
-  if (isLoading) {
-    return <p>Loading...</p>;
-  }
 
   return (
     <>
